pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "label_pizza"
version = "0.1.0"
description = "Video labeling platform"
authors = [
    { name = "Zhiqiu Lin", email = "zl279@cornell.edu" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy",
    "supabase",
    "streamlit",
    "streamlit-feedback",
    "opencv-python-headless",
    "numpy",
    "pandas",
    "matplotlib",
    "psycopg[binary]",
    "python-dotenv",
    "pytest",
]

# This section tells setuptools to find packages automatically
# It will find the 'label_pizza' directory as your main package
[tool.setuptools.packages.find]
exclude = ["tests*"]  # Exclude test directories